+++ src/DataFixtures/AppFixtures.php
@@ -9,6 +9,7 @@ use App\Entity\Project;
 use App\Entity\Version;
 use App\Entity\Worker;
 use App\Entity\Worklog;
+use App\Enum\BillableKindsEnum;
 use App\Enum\ClientTypeEnum;
 use App\Service\JiraApiService;
 use App\Service\LeantimeApiService;
@@ -70,6 +71,8 @@ class AppFixtures extends Fixture
             }
 
             for ($i = 0; $i < 10; ++$i) {
+                $modBillable = 0 == $i % 2 ? true : false;
+
                 $project = new Project();
                 $project->setName("project-$key-$i");
                 $project->setProjectTrackerId("project-$key-$i");
@@ -79,6 +82,7 @@ class AppFixtures extends Fixture
                 $project->setProjectLeadMail('test@economics.local.itkdev.dk');
                 $project->setProjectLeadName('Test Testesen');
                 $project->setDataProvider($dataProvider);
+                $project->setIsBillable($modBillable);
 
                 $manager->persist($project);
 
@@ -122,18 +126,26 @@ class AppFixtures extends Fixture
                     $manager->persist($issue);
 
                     for ($k = 0; $k < 100; ++$k) {
+                        $year = (new \DateTime())->format('Y');
+
+                        // Use modulo to get months and dates to create started-dates spanning the entire year
+                        $modMonth = str_pad((string) ($k % 12 + 1), 2, '0', STR_PAD_LEFT);
+                        $modDay = str_pad((string) ($k % 28 + 1), 2, '0', STR_PAD_LEFT);
+
+                        $modKind = 0 == $i % 2 ? BillableKindsEnum::GENERAL_BILLABLE : null;
+
                         $worklog = new Worklog();
                         $worklog->setProjectTrackerIssueId("worklog-$key-$i-$j-$k");
                         $worklog->setWorklogId($i * 100000 + $j * 1000 + $k);
                         $worklog->setDescription("Beskrivelse af worklog-$key-$i-$j-$k");
                         $worklog->setIsBilled(false);
                         $worklog->setProject($project);
-                        $worklog->setWorker($workerArray[rand(0, 9)]);
+                        $worklog->setWorker($workerArray[$i % 10]);
                         $worklog->setTimeSpentSeconds(60 * 15 * ($k + 1));
-                        $worklog->setStarted(\DateTime::createFromFormat('U', (string) rand(strtotime(date('Y-01-01')), strtotime(date('Y-12-31')))));
+                        $worklog->setStarted(\DateTime::createFromFormat('U', (string) strtotime("$year-$modMonth-$modDay"), new \DateTimeZone('Europe/Copenhagen')));
                         $worklog->setIssue($issue);
                         $worklog->setDataProvider($dataProvider);
-
+                        $worklog->setKind($modKind);
                         $manager->persist($worklog);
                     }