+++ src/Model/Invoices/ProjectFilterData.php
@@ -7,4 +7,5 @@ class ProjectFilterData
     public ?string $name = null;
     public ?string $key = null;
     public ?bool $include = null;
+    public ?bool $isBillable = null;
 }