+++ src/Model/Reports/WorkloadReportFormData.php
@@ -7,5 +7,6 @@ use App\Entity\DataProvider;
 class WorkloadReportFormData
 {
     public DataProvider $dataProvider;
-    public string $viewMode;
+    public WorkloadReportPeriodTypeEnum $viewPeriodType;
+    public WorkloadReportViewModeEnum $viewMode;
 }