+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   NOTE: APP_DEFAULT_PLANNING_DATA_PROVIDER has been changed to APP_DEFAULT_DATA_PROVIDER. This has to be changed when releasing.
 * [PR-124](https://github.com/itk-dev/economics/pull/124)
   710: Added workload report
+* [PR-129](https://github.com/itk-dev/economics/pull/129)
+  1632: Added invoicing rate view to workload report
 
 ## [2.3.3] - 2024-07-10